+++ pandas/core/indexes/datetimelike.py
@@ -367,76 +367,6 @@ class DatetimeIndexOpsMixin(DatetimeLikeArrayMixin):
 
     contains = __contains__
 
-    def __getitem__(self, key):
-        """
-        This getitem defers to the underlying array, which by-definition can
-        only handle list-likes, slices, and integer scalars
-        """
-
-        is_int = is_integer(key)
-        if is_scalar(key) and not is_int:
-            raise IndexError("only integers, slices (`:`), ellipsis (`...`), "
-                             "numpy.newaxis (`None`) and integer or boolean "
-                             "arrays are valid indices")
-
-        getitem = self._data.__getitem__
-        if is_int:
-            val = getitem(key)
-            return self._box_func(val)
-        else:
-            if com.is_bool_indexer(key):
-                key = np.asarray(key)
-                if key.all():
-                    key = slice(0, None, None)
-                else:
-                    key = lib.maybe_booleans_to_slice(key.view(np.uint8))
-
-            attribs = self._get_attributes_dict()
-
-            is_period = isinstance(self, ABCPeriodIndex)
-            if is_period:
-                freq = self.freq
-            else:
-                freq = None
-                if isinstance(key, slice):
-                    if self.freq is not None and key.step is not None:
-                        freq = key.step * self.freq
-                    else:
-                        freq = self.freq
-
-            attribs['freq'] = freq
-
-            result = getitem(key)
-            if result.ndim > 1:
-                # To support MPL which performs slicing with 2 dim
-                # even though it only has 1 dim by definition
-                if is_period:
-                    return self._simple_new(result, **attribs)
-                return result
-
-            return self._simple_new(result, **attribs)
-
-    def _nat_new(self, box=True):
-        """
-        Return Index or ndarray filled with NaT which has the same
-        length as the caller.
-
-        Parameters
-        ----------
-        box : boolean, default True
-            - If True returns a Index as the same as caller.
-            - If False returns ndarray of np.int64.
-        """
-        result = np.zeros(len(self), dtype=np.int64)
-        result.fill(iNaT)
-        if not box:
-            return result
-
-        attribs = self._get_attributes_dict()
-        if not is_period_dtype(self):
-            attribs['freq'] = None
-        return self._simple_new(result, **attribs)
-
     # Try to run function on index first, and then on elements of index
     # Especially important for group-by functionality
     def map(self, f):
@@ -669,17 +599,6 @@ class DatetimeIndexOpsMixin(DatetimeLikeArrayMixin):
         return (super(DatetimeIndexOpsMixin, self)
                 ._convert_scalar_indexer(key, kind=kind))
 
-    def _add_nat(self):
-        """Add pd.NaT to self"""
-        if is_period_dtype(self):
-            raise TypeError('Cannot add {cls} and {typ}'
-                            .format(cls=type(self).__name__,
-                                    typ=type(NaT).__name__))
-
-        # GH#19124 pd.NaT is treated like a timedelta for both timedelta
-        # and datetime dtypes
-        return self._nat_new(box=True)
-
     def _addsub_offset_array(self, other, op):
         """
         Add or subtract array-like of DateOffset objects
@@ -706,7 +625,7 @@ class DatetimeIndexOpsMixin(DatetimeLikeArrayMixin):
         kwargs = {}
         if not is_period_dtype(self):
             kwargs['freq'] = 'infer'
-        return self._constructor(res_values, **kwargs)
+        return type(self)(res_values, **kwargs)
 
     @classmethod
     def _add_datetimelike_methods(cls):
